Created seperate file for the name tag trait.
