+++ src/main/java/de/craften/plugins/rpgplus/components/entitymanager/RpgPlusEntity.java
@@ -2,20 +2,15 @@ package de.craften.plugins.rpgplus.components.entitymanager;
 
 import net.citizensnpcs.api.CitizensAPI;
 import net.citizensnpcs.api.ai.TargetType;
-import net.citizensnpcs.api.event.NPCDeathEvent;
 import net.citizensnpcs.api.npc.NPC;
-import net.citizensnpcs.api.trait.Trait;
 
 import org.bukkit.Location;
-import org.bukkit.entity.ArmorStand;
-import org.bukkit.entity.Creature;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.LivingEntity;
-import org.bukkit.event.EventHandler;
 import org.bukkit.event.player.PlayerTeleportEvent;
 
-import de.craften.plugins.rpgplus.util.EntityUtil;
+import de.craften.plugins.rpgplus.components.entitymanager.traits.NameTagTrait;
 
 /**
  * A basic managed entity without any special logic.
@@ -27,16 +22,12 @@ public class RpgPlusEntity<T extends Entity> {
     
     private EntityType type;
     
-    private boolean isNameVisible;
-    private String name;
-    private String secondName;
-    private ArmorStand nametag;
-    private ArmorStand secondNameTag;
-    
     public RpgPlusEntity(Location location, EntityType type) {
         this.location = location;
         this.type = type;
         npc = CitizensAPI.getNPCRegistry().createNPC(type, "");
+
+        npc.addTrait(new NameTagTrait(type));
         
         if (type == EntityType.PLAYER) {
         	npc.data().set(NPC.NAMEPLATE_VISIBLE_METADATA, true);
@@ -49,94 +40,11 @@ public class RpgPlusEntity<T extends Entity> {
     public T spawn() {
         // TODO support custom skins
         // npc.data().setPersistent(NPC.PLAYER_SKIN_UUID_METADATA, "leMaik");
-
-        npc.addTrait(new Trait("NameTagTrait") {
-        	@Override
-        	public void run() {
-        		if (nametag != null) {
-                   
-        			nametag.teleport(npc.getEntity().getLocation().clone().add(0, EntityUtil.getEntityHeight(npc.getEntity()), 0));
-        		}
-        		if (secondNameTag != null) {
-
-                    secondNameTag.teleport(npc.getEntity().getLocation().clone().add(0, EntityUtil.getEntityHeight(npc.getEntity()) + EntityUtil.NAME_TAG_HEIGHT + 0.1, 0));
-        		}
-        	}
-        	
-        	@Override
-        	public void onRemove() {
-                if (nametag != null) {
-                    nametag.remove();
-                }
-                if (secondNameTag != null) {
-                    secondNameTag.remove();
-                }
-        	}
-        	
-        	@Override
-        	public void onDespawn() {
-                if (nametag != null) {
-                    nametag.remove();
-                }
-                if (secondNameTag != null) {
-                    secondNameTag.remove();
-                }
-        	}
-        	
-        	@EventHandler
-        	public void onDeath(NPCDeathEvent event) {
-        		if (event.getNPC().getUniqueId().equals(npc.getUniqueId())) {
-        			if (nametag != null) {
-                        nametag.remove();
-                    }
-                    if (secondNameTag != null) {
-                        secondNameTag.remove();
-                    }
-        		}
-        	}
-        });
-        
+    	
         npc.spawn(location);
-
+        
         if (npc.getEntity() instanceof LivingEntity && !isTakingDamage) {
             npc.setProtected(true);
-        }
-        
-
-        if (type != EntityType.PLAYER) {
-        	if (npc.getEntity() instanceof LivingEntity && !name.isEmpty() && isNameVisible) {
-        		Location nameTagLocation = npc.getEntity().getLocation().clone().add(0, EntityUtil.getEntityHeight(npc.getEntity()), 0);
-        		
-            	if (nametag == null) {
-            		nametag = npc.getEntity().getWorld().spawn(nameTagLocation, ArmorStand.class);
-            	}
-            	
-            	nametag.teleport(nameTagLocation);
-            	nametag.setCustomName(name);
-            	nametag.setCustomNameVisible(true);
-            	nametag.setVisible(false);
-            	nametag.setMarker(true);
-            	nametag.setCanPickupItems(false);
-            	nametag.setGravity(false);
-            	
-        	}
-        }
-        
-        if (npc.getEntity() != null && npc.getEntity() instanceof LivingEntity && !secondName.isEmpty() && isNameVisible) {
-            Location nameTagLocation = npc.getEntity().getLocation().clone().add(0, EntityUtil.getEntityHeight(npc.getEntity()) + EntityUtil.NAME_TAG_HEIGHT + 0.1, 0);
-            
-            if (secondNameTag == null) {
-                secondNameTag = npc.getEntity().getWorld().spawn(nameTagLocation, ArmorStand.class);
-            }
-
-            secondNameTag.teleport(nameTagLocation);
-            secondNameTag.setCustomName(secondName);
-            secondNameTag.setCustomNameVisible(true);
-            secondNameTag.setVisible(false);
-            secondNameTag.setMarker(true);
-            secondNameTag.setCanPickupItems(false);
-            secondNameTag.setGravity(false);
-
         }
         
         return (T) npc.getEntity();
@@ -147,29 +55,22 @@ public class RpgPlusEntity<T extends Entity> {
     }
 
     public String getName() {
-        return name;
+        return getNameTagTrait().getFirstName();
     }
 
     public void setName(String name) {
-    	this.name = name;
+    	getNameTagTrait().setFirstName(name);
     	if (type == EntityType.PLAYER) {
     	   npc.setName(name);
-       } else {
-    	   if (nametag != null) {
-    		   nametag.setCustomName(name);
-    	   }
-       }
+    	}
     }
 
     public String getSecondName() {
-        return secondName;
+        return getNameTagTrait().getSecondName();
     }
 
     public void setSecondName(String secondName) {
-        this.secondName = secondName;
-        if (secondNameTag != null) {
-        	secondNameTag.setCustomName(name);
- 	   	}
+    	getNameTagTrait().setSecondName(secondName);
     }
     
     public boolean isTakingDamage() {
@@ -182,11 +83,11 @@ public class RpgPlusEntity<T extends Entity> {
     }
 
     public boolean isNameVisible() {
-        return isNameVisible;
+        return getNameTagTrait().isVisible();
     }
 
     public void setNameVisible(boolean nameVisible) {
-       this.isNameVisible = nameVisible;
+      	getNameTagTrait().setVisible(nameVisible);
     }
 
     /**
@@ -225,5 +126,9 @@ public class RpgPlusEntity<T extends Entity> {
             return (T) npc.getEntity();
         }
         return null;
+    }
+    
+    public NameTagTrait getNameTagTrait() {
+        return npc.getTrait(NameTagTrait.class);
     }
 }
